Adds select/deselect context actions
